slugs: recompute slugs if one of them changed
When changing a file title, the slug in the URL should also change.
