+++ src/fava/application.py
@@ -18,6 +18,7 @@ from io import BytesIO
 from typing import Any
 from typing import Dict
 from typing import Optional
+from typing import List
 
 import flask
 import markdown2  # type: ignore
@@ -45,6 +46,7 @@ from fava.help import HELP_PAGES
 from fava.helpers import FavaAPIException
 from fava.json_api import json_api
 from fava.serialisation import serialise
+from fava.util import next_key
 from fava.util import resource_path
 from fava.util import send_file_inline
 from fava.util import setup_logging
@@ -91,19 +93,31 @@ REPORTS = [
 LOAD_FILE_LOCK = threading.Lock()
 
 
-def _load_file():
+def ledger_slug(ledger: FavaLedger) -> str:
+    """Generate URL slug for a ledger."""
+    title_slug = slugify(ledger.options["title"])
+    return title_slug or slugify(ledger.beancount_file_path)
+
+
+def update_ledger_slugs(ledgers: List[FavaLedger]) -> None:
+    """Update the dictionary mapping URL slugs to ledgers."""
+    ledgers_by_slug: Dict[str, FavaLedger] = {}
+    for ledger in ledgers:
+        slug = ledger_slug(ledger)
+        unique_key = next_key(slug, ledgers_by_slug)
+        ledgers_by_slug[unique_key] = ledger
+    app.config["LEDGERS"] = ledgers_by_slug
+
+
+def _load_file() -> None:
     """Load Beancount files.
 
     This is run automatically on the first request.
     """
-    app.config["LEDGERS"] = {}
-    for filepath in app.config["BEANCOUNT_FILES"]:
-        ledger = FavaLedger(filepath)
-        slug = slugify(ledger.options["title"])
-        if not slug:
-            slug = slugify(filepath)
-        app.config["LEDGERS"][slug] = ledger
-    app.config["FILE_SLUGS"] = list(app.config["LEDGERS"].keys())
+    ledgers = [
+        FavaLedger(filepath) for filepath in app.config["BEANCOUNT_FILES"]
+    ]
+    update_ledger_slugs(ledgers)
 
 
 BABEL = Babel(app)
@@ -229,8 +243,14 @@ def _pull_beancount_file(_, values) -> None:
         if not app.config.get("LEDGERS"):
             _load_file()
     if g.beancount_file_slug:
-        if g.beancount_file_slug not in app.config["FILE_SLUGS"]:
-            abort(404)
+        if g.beancount_file_slug not in app.config["LEDGERS"]:
+            if not any(
+                g.beancount_file_slug == ledger_slug(ledger)
+                for ledger in app.config["LEDGERS"].values()
+            ):
+                abort(404)
+            # one of the file slugs changed, update the mapping
+            update_ledger_slugs(app.config["LEDGERS"].values())
         g.ledger = app.config["LEDGERS"][g.beancount_file_slug]
         g.conversion = request.args.get(
             "conversion", g.ledger.fava_options["conversion"]
@@ -253,7 +273,7 @@ def fava_api_exception(error: FavaAPIException):
 def index():
     """Redirect to the Income Statement (of the given or first file)."""
     if not g.beancount_file_slug:
-        g.beancount_file_slug = app.config["FILE_SLUGS"][0]
+        g.beancount_file_slug = next(iter(app.config["LEDGERS"]))
     return redirect(url_for("report", report_name="income_statement"))
